+++ lib/console/filter.rb
@@ -11,7 +11,7 @@ module Console
 	
 	# A log filter which can be used to filter log messages based on severity, subject, and other criteria.
 	class Filter
-		if Object.const_defined?(:Ractor) and RUBY_VERSION >= "3.1"
+		if Object.const_defined?(:Ractor) and RUBY_VERSION >= "3.4"
 			# Define a method which can be shared between ractors.
 			def self.define_immutable_method(name, &block)
 				block = Ractor.make_shareable(block)