+++ releases.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
   - Add `fiber_id` to serialized output records to help identify which fiber logged the message.
+  - Ractor support appears broken in older Ruby versions, so we now require Ruby 3.4 or later for Ractor compatibility, if you need Ractor support.
 
 ## v1.31.0